Marking EndSession as deprecated
It is not only not really useful in blueprints, but people have also been using it wrong.

It isn't useful to Steam Subsystem and isn't really useful in blueprints to begin with as its use is vauge, would rather enforce the actual
total destruction of the session instead by making users call DestroySession as is intended.

I feel like that node suffers from overexposing the back end to blueprints.


Former-commit-id: 77ce18bce6eb7f400ca744ef279cc8bba91f7073